pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pipeworks-namegen-lexicon"
version = "0.1.1"
description = "Pipeworks Namegen Lexicon pipeline and web tooling"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "GNU General Public License v3.0"}
authors = [{name = "Pipe-Works"}]
dependencies = [
  "pipeworks-ipc @ git+https://github.com/pipe-works/pipeworks_ipc.git@pipeworks-ipc-v0.1.3",
  "pyyaml>=6.0",
  "numpy>=1.24.0",
]

[project.scripts]
pipeworks-namegen-lexicon-web = "build_tools.syllable_walk_web.cli:main"

[project.optional-dependencies]
dev = [
  "black>=24.0",
  "bandit[toml]>=1.7",
  "cmudict>=1.0.0",
  "langdetect>=1.0.9",
  "matplotlib>=3.7.0",
  "mypy>=1.10",
  "pandas>=2.0.0",
  "plotly>=5.0.0",
  "pre-commit>=3.5",
  "pyphen>=0.14.0",
  "pytest>=8.0",
  "pytest-asyncio>=0.23.0",
  "pytest-cov>=5.0",
  "scikit-learn>=1.3.0",
  "types-pyyaml>=6.0",
  "ruff>=0.5",
]
docs = [
  "sphinx>=8.0,<9.0",
  "sphinx-argparse>=0.4.0",
  "sphinx-autodoc-typehints>=3.5",
  "sphinx-autoapi>=3.0",
  "sphinx-rtd-theme>=3.0",
  "myst-parser>=4.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"build_tools.syllable_walk_web" = [
  "schemas/*.json",
  "static/*",
]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"

[tool.mypy]
python_version = "3.12"
mypy_path = ["src"]

[[tool.mypy.overrides]]
module = "cmudict.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pipeworks_ipc.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "numpy.*"
ignore_missing_imports = true

[tool.bandit]
skips = ["B101", "B108", "B311", "B608"]
